Passkey refactored
